finished project

+++ frontend/src/app/admin/views/companies-view/delete-company/delete-company.component.ts
@@ -1,7 +1,8 @@
 import {Component, Input} from '@angular/core';
 import {finalize} from "rxjs";
-import { Company } from 'src/app/models/company.model';
-import { CompanyService } from 'src/app/services/company.service';
+import {ConfirmationService} from "primeng/api";
+import {Company} from 'src/app/models/company.model';
+import {CompanyService} from 'src/app/services/company.service';
 
 @Component({
   selector: 'app-delete-company',
@@ -13,16 +14,24 @@ export class DeleteCompanyComponent {
   @Input() company: Company;
   loading: boolean;
 
-  constructor(private companyService: CompanyService) {
+  constructor(private companyService: CompanyService,
+              private confirmationService: ConfirmationService) {
   }
 
   delete() {
-    this.loading = true;
-    this.companyService.delete(this.company.id)
-      .pipe(finalize(() => this.loading = false))
-      .subscribe({
-        next: () => {
-        }
-      })
+    this.confirmationService.confirm({
+      header: 'Kunde löschen',
+      message: `Möchtest du "${this.company.name}" wirklich löschen?`,
+      icon: 'pi pi-exclamation-triangle',
+      accept: () => {
+        this.loading = true;
+        this.companyService.delete(this.company.id)
+          .pipe(finalize(() => this.loading = false))
+          .subscribe({
+            next: () => {
+            }
+          })
+      }
+    });
   }
 }
