+++ frontend/src/app/admin/views/project-category-view/delete-project-category/delete-project-category.component.ts
@@ -1,8 +1,9 @@
 import {Component, Input} from '@angular/core';
 import {finalize} from "rxjs";
+import {ConfirmationService} from "primeng/api";
 
-import { ProjectType } from 'src/app/models/project-type.model';
-import { ProjectTypeService } from 'src/app/services/project-type.service';
+import {ProjectType} from 'src/app/models/project-type.model';
+import {ProjectTypeService} from 'src/app/services/project-type.service';
 
 @Component({
   selector: 'app-delete-project-category',
@@ -14,16 +15,24 @@ export class DeleteProjectCategoryComponent {
   @Input() projectType: ProjectType
   loading: boolean;
 
-  constructor(private projectTypeService: ProjectTypeService) {
+  constructor(private projectTypeService: ProjectTypeService,
+              private confirmationService: ConfirmationService) {
   }
 
   delete() {
-    this.loading = true;
-    this.projectTypeService.delete(this.projectType.id)
-      .pipe(finalize(() => this.loading = false))
-      .subscribe({
-        next: () => {
-        }
-      })
+    this.confirmationService.confirm({
+      header: 'Projektart löschen',
+      message: `Möchtest du "${this.projectType.name}" wirklich löschen?`,
+      icon: 'pi pi-exclamation-triangle',
+      accept: () => {
+        this.loading = true;
+        this.projectTypeService.delete(this.projectType.id)
+          .pipe(finalize(() => this.loading = false))
+          .subscribe({
+            next: () => {
+            }
+          })
+      }
+    });
   }
 }
